requirements.txt
torch
torchvision
transformers
matplotlib
numpy
pillow
scikit-learn
tqdm
